requirements.txt
google-adk>=1.27.0
google-genai>=1.0.0
google-cloud-firestore>=2.19.0
google-cloud-aiplatform>=1.74.0
a2a-sdk>=0.3.0
httpx>=0.27.0
youtube-transcript-api>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
